requirements.txt
langchain>=0.1.0
groq>=0.4.0
msal>=1.25.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0